Adds the first three endpoints per the spec:  * GET on /id will return the stream  * POST on /id will create the stream (not upsert)  * PUT on /id will update the stream (not upsert)
Caveats: the actual file handling is stubbed right now, filters (encryption, compression) are not in place. The latter should probably be handled through Accept/Content-Encoding headers in a non-pedagogical situation.

Streams are handled via application/octet-stream instead of the multipart that the spec seemed to call for. The latter didn't really make sense to me - though can certainly be added if actually required. The pattern for valid IDs was taken from what Amazon's S3 suggests as it's familiar to most people.

Integration phase is not bound yet, neither does Jetty run for the duration.